requirements.txt
flask
celery
redis
yt-dlp
numpy
opencv-python-headless
moviepy
openai-whisper
